backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
pydantic==2.5.0
bcrypt==4.1.1
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.23
PyJWT==2.11.0
email-validator==2.1.0
